requirements.txt

fastapi
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
pymupdf==1.24.10
langdetect
regex==2024.9.11
phonenumbers==8.13.47
tldextract==5.1.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.34
qdrant-client==1.9.2
sentence-transformers
torch==2.3.1
numpy==1.26.4
pandas==2.2.2
neo4j==5.23.1
langchain
langgraph
networkx
orjson
